+++ Ann_core/src/ann/__init__.py
@@ -1,3 +1,3 @@
 """Ann Core package."""
 
-__version__ = "0.0.11"
+__version__ = "0.0.12"